+++ src/main/java/com/github/yury1991/PayrollWeb/models/Rate.java
@@ -2,5 +2,5 @@ package com.github.yury1991.PayrollWeb.models;
 
 //Интерфейс коэффициента
 public interface Rate {
-	public double getRate();
+	public float getRate();
 }
